Split out tests for Enum and Variable types
