+++ tsconfig.json
@@ -8,6 +8,7 @@
         "sourceMap": true,
         "experimentalDecorators": true,
         "removeComments": true,
+        "noUnusedLocals": true,
         "lib": [
             "es2015"
         ]
